dom/tests/mochitest/mouselock/test_userPref.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id={633602}
  -->
<head>
  <title>Test for Bug {633602}</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js" type="text/javascript"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id={BUGNUMBER}">Mozilla Bug 633602</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

  <canvas id="canvas" width="150" height="150"></canvas>

  <pre id="test">
  <script class="testbody" type="text/javascript">
    SimpleTest.waitForExplicitFinish();

    var pointer = navigator.pointer;
    var canvas = document.getElementById("canvas");
    var totalTests = 2;
    var nTests = 0;
    SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);

    document.addEventListener("mozfullscreenchange", function (e) {
      if (document.mozFullScreen) {
        if (document.mozFullScreenElement === canvas) {
        // Set Preference to False to see if lock fails
          SpecialPowers.setBoolPref("full-screen-api.mouse-lock.enabled", false);
          pointer.lock(canvas);
          nTests += 1;
          is(pointer.islocked(), false, "Mouse shouldn't be locked when pref setting is false.");

          // Set Preference to True to see if lock succeeds 
          SpecialPowers.setBoolPref("full-screen-api.mouse-lock.enabled", true);
          pointer.lock(canvas);
          nTests += 1;
          is(pointer.islocked(), true, "Mouse should be locked when pref setting is true.");
        }
         
        if(nTests === totalTests) {
          document.mozCancelFullScreen();
          SimpleTest.finish();
        }
      }
    }, false);
    
    canvas.mozRequestFullScreen(); 
    setTimeout("document.body.mozRequestFullScreen();", 300);
  </script>
  </pre>
</body>
</html>
